WIP: started ripping out the async_await feature flag
