+++ lib/config/index.js
@@ -859,13 +859,6 @@ Config.prototype.logUnknown = function logUnknown(json, key) {
 }
 
 /**
- * Return the availability of async_hook for use by the agent.
- */
-Config.prototype.checkAsyncHookStatus = function checkAsyncHookStatus() {
-  return this.feature_flag.await_support
-}
-
-/**
  * Gets the user set host display name. If not provided, it returns the default value.
  *
  * This function is written is this strange way becauase of the use of caching variables.
