+++ lib/feature_flags.js
@@ -7,7 +7,6 @@
 
 // unreleased flags gating an active feature
 exports.prerelease = {
-  await_support: true,
   certificate_bundle: false,
   express5: false,
   new_promise_tracking: false,
@@ -31,7 +30,8 @@ exports.released = [
   'send_request_uri_attribute',
   'synthetics',
   'dt_format_w3c',
-  'fastify_instrumentation'
+  'fastify_instrumentation',
+  'await_support'
 ]
 
 // flags that are no longer used for unreleased features
