+++ lib/instrumentation/core/async_hooks.js
@@ -6,48 +6,11 @@
 'use strict'
 
 const logger = require('../../logger').child({ component: 'async_hooks' })
-const promInit = require('../promise')
+const asyncHooks = require('async_hooks')
 
 module.exports = initialize
 
-/**
- * The spec for the native `Promise` class.
- */
-const STATIC_PROMISE_METHODS = ['accept', 'all', 'defer', 'race', 'reject', 'resolve']
-const NATIVE_PROMISE_SPEC = {
-  name: 'global',
-  constructor: 'Promise',
-  executor: true,
-  useFinally: false,
-  $proto: {
-    then: ['then', 'chain'],
-    catch: ['catch']
-  },
-  $static: {
-    $copy: STATIC_PROMISE_METHODS,
-    cast: STATIC_PROMISE_METHODS
-  }
-}
-
 function initialize(agent, shim) {
-  const enableHooks = agent.config.checkAsyncHookStatus()
-  if (enableHooks && tryAsyncHooks(agent, shim)) {
-    logger.debug('Using async_hooks.')
-  } else {
-    logger.debug('Using promise instrumentation.')
-    promInit(agent, global, NATIVE_PROMISE_SPEC)
-  }
-}
-
-function tryAsyncHooks(agent, shim) {
-  let asyncHooks = null
-  try {
-    asyncHooks = require('async_hooks')
-  } catch (e) {
-    logger.info(e, 'Not using async_hooks module.')
-    return false
-  }
-
   // this map is reused to track the segment that was active when
   // the before callback is called to be replaced in the after callback
   const segmentMap = new Map()
