--- test/integration/core/promises_hook.tap.js
@@ -1,8 +1,0 @@
-/*
- * Copyright 2020 New Relic Corporation. All rights reserved.
- * SPDX-License-Identifier: Apache-2.0
- */
-
-'use strict'
-
-require('./promises.js')({ await_support: true })
